+++ src/os/win32/ngx_files.c
@@ -73,7 +73,7 @@ ngx_read_file(ngx_file_t *file, u_char *buf, size_t size, off_t offset)
 
     povlp = &ovlp;
 
-    if (ReadFile(file->fd, buf, size, &n, povlp) == 0) {
+    if (ReadFile(file->fd, buf, (DWORD)size, &n, povlp) == 0) {
         err = ngx_errno;
 
         if (err == ERROR_HANDLE_EOF) {
@@ -105,7 +105,7 @@ ngx_write_file(ngx_file_t *file, u_char *buf, size_t size, off_t offset)
 
     povlp = &ovlp;
 
-    if (WriteFile(file->fd, buf, size, &n, povlp) == 0) {
+    if (WriteFile(file->fd, buf,(DWORD) size, &n, povlp) == 0) {
         ngx_log_error(NGX_LOG_ERR, file->log, ngx_errno,
                       "WriteFile() \"%s\" failed", file->name.data);
         return NGX_ERROR;
@@ -166,7 +166,7 @@ ngx_read_fd(ngx_fd_t fd, void *buf, size_t size)
 {
     u_long  n;
 
-    if (ReadFile(fd, buf, size, &n, NULL) != 0) {
+    if (ReadFile(fd, buf, (DWORD)size, &n, NULL) != 0) {
         return (size_t) n;
     }
 
@@ -179,7 +179,7 @@ ngx_write_fd(ngx_fd_t fd, void *buf, size_t size)
 {
     u_long  n;
 
-    if (WriteFile(fd, buf, size, &n, NULL) != 0) {
+    if (WriteFile(fd, buf,(DWORD) size, &n, NULL) != 0) {
         return (size_t) n;
     }
 
@@ -192,9 +192,9 @@ ngx_write_console(ngx_fd_t fd, void *buf, size_t size)
 {
     u_long  n;
 
-    (void) CharToOemBuff(buf, buf, size);
+    (void) CharToOemBuff(buf, buf, (DWORD)size);
 
-    if (WriteFile(fd, buf, size, &n, NULL) != 0) {
+    if (WriteFile(fd, buf, (DWORD)size, &n, NULL) != 0) {
         return (size_t) n;
     }
 
@@ -756,7 +756,7 @@ ngx_win32_check_filename(u_char *name, u_short *u, size_t len)
         return NGX_ERROR;
     }
 
-    n = GetLongPathNameW(u, lu, len);
+    n = GetLongPathNameW(u, lu, (DWORD)len);
 
     if (n == 0) {
         goto failed;
